step1 - 파일 분리 및 하위 클래스 생성
